// src/app/module/Course/course.model.ts
import { model, Schema} from "mongoose";
import { TCourse, TCourseFaculty, TPreRequisiteCourses } from "./course.interface";

const preRequisiteCoursesSchema = new Schema<TPreRequisiteCourses>({
    course:{
        type: Schema.Types.ObjectId,
        ref: "Course"
    },
    isDeleted:{
        type:Boolean,
        default: false,
    }
})

const courseSchema = new Schema<TCourse>({
    title:{
        type:String,
        trim: true,
        unique: true,
        required: [true, "Title is Required!!!"],
    },
    prefix:{
        type: String,
        trim: true,
        required: true,
    },
    code:{
        type: Number,
        trim: true,
        required: true,
    },
    credit:{
        type: Number,
        trim: true,
        required: true,
    },
    preRequisiteCourses: [ preRequisiteCoursesSchema ],
    isDeleted: {
        type: Boolean,
        default: false,
      },
});

export const Course = model<TCourse>('Course',courseSchema)



const courseFacultySchema = new Schema<TCourseFaculty>({
    course:{
        type: Schema.Types.ObjectId,
        ref: "Course",
        unique: true,
    },
    faculties: {
        type: [{
            type: Schema.Types.ObjectId,
            ref: "Faculty",

        }]
    }
});

export const CourseFaculty = model<TCourseFaculty>('CourseFaculty',courseFacultySchema)

